Introduce `AbstractMemory`
