+++ interpreter/src/main/java/org/qbicc/interpreter/memory/AbstractMemory.java
@@ -1,0 +1,114 @@
+package org.qbicc.interpreter.memory;
+
+import org.qbicc.graph.atomic.ReadAccessMode;
+import org.qbicc.graph.atomic.WriteAccessMode;
+import org.qbicc.interpreter.Memory;
+import org.qbicc.interpreter.VmObject;
+import org.qbicc.interpreter.impl.InvalidMemoryAccessException;
+import org.qbicc.pointer.Pointer;
+import org.qbicc.type.ValueType;
+
+/**
+ *
+ */
+public abstract class AbstractMemory implements Memory {
+    // load
+
+    public int load8(long index, ReadAccessMode mode) {
+        throw new InvalidMemoryAccessException();
+    }
+
+    public int load16(long index, ReadAccessMode mode) {
+        throw new InvalidMemoryAccessException();
+    }
+
+    public int load32(long index, ReadAccessMode mode) {
+        throw new InvalidMemoryAccessException();
+    }
+
+    public long load64(long index, ReadAccessMode mode) {
+        throw new InvalidMemoryAccessException();
+    }
+
+    public VmObject loadRef(long index, ReadAccessMode mode) {
+        throw new InvalidMemoryAccessException();
+    }
+
+    public ValueType loadType(long index, ReadAccessMode mode) {
+        throw new InvalidMemoryAccessException();
+    }
+
+    public Pointer loadPointer(long index, ReadAccessMode mode) {
+        throw new InvalidMemoryAccessException();
+    }
+
+    // store
+
+    public void store8(long index, int value, WriteAccessMode mode) {
+        throw new InvalidMemoryAccessException();
+    }
+
+    public void store16(long index, int value, WriteAccessMode mode) {
+        throw new InvalidMemoryAccessException();
+    }
+
+    public void store32(long index, int value, WriteAccessMode mode) {
+        throw new InvalidMemoryAccessException();
+    }
+
+    public void store64(long index, long value, WriteAccessMode mode) {
+        throw new InvalidMemoryAccessException();
+    }
+
+    public void storeRef(long index, VmObject value, WriteAccessMode mode) {
+        throw new InvalidMemoryAccessException();
+    }
+
+    public void storeType(long index, ValueType value, WriteAccessMode mode) {
+        throw new InvalidMemoryAccessException();
+    }
+
+    public void storePointer(long index, Pointer value, WriteAccessMode mode) {
+        throw new InvalidMemoryAccessException();
+    }
+
+    // atomic CAS
+
+    @Override
+    public int compareAndExchange8(long index, int expect, int update, ReadAccessMode readMode, WriteAccessMode writeMode) {
+        throw new InvalidMemoryAccessException();
+    }
+
+    @Override
+    public int compareAndExchange16(long index, int expect, int update, ReadAccessMode readMode, WriteAccessMode writeMode) {
+        throw new InvalidMemoryAccessException();
+    }
+
+    @Override
+    public int compareAndExchange32(long index, int expect, int update, ReadAccessMode readMode, WriteAccessMode writeMode) {
+        throw new InvalidMemoryAccessException();
+    }
+
+    @Override
+    public long compareAndExchange64(long index, long expect, long update, ReadAccessMode readMode, WriteAccessMode writeMode) {
+        throw new InvalidMemoryAccessException();
+    }
+
+    @Override
+    public VmObject compareAndExchangeRef(long index, VmObject expect, VmObject update, ReadAccessMode readMode, WriteAccessMode writeMode) {
+        throw new InvalidMemoryAccessException();
+    }
+
+    @Override
+    public ValueType compareAndExchangeType(long index, ValueType expect, ValueType update, ReadAccessMode readMode, WriteAccessMode writeMode) {
+        throw new InvalidMemoryAccessException();
+    }
+
+    @Override
+    public Pointer compareAndExchangePointer(long index, Pointer expect, Pointer update, ReadAccessMode readMode, WriteAccessMode writeMode) {
+        throw new InvalidMemoryAccessException();
+    }
+
+    @Override
+    public abstract Memory clone();
+}
