+++ interpreter/src/main/java/org/qbicc/interpreter/memory/CharArrayMemory.java
@@ -13,15 +13,11 @@ import java.util.Arrays;
 import org.qbicc.graph.atomic.ReadAccessMode;
 import org.qbicc.graph.atomic.WriteAccessMode;
 import org.qbicc.interpreter.Memory;
-import org.qbicc.interpreter.VmObject;
-import org.qbicc.interpreter.impl.InvalidMemoryAccessException;
-import org.qbicc.pointer.Pointer;
-import org.qbicc.type.ValueType;
 
 /**
  * A memory region which is backed by a {@code char} array which can be directly accessed.
  */
-public final class CharArrayMemory implements Memory {
+public final class CharArrayMemory extends AbstractMemory {
     private static final VarHandle h16 = ConstantBootstraps.arrayVarHandle(MethodHandles.lookup(), "ignored", VarHandle.class, char[].class);
 
     private final char[] array;
@@ -32,11 +28,6 @@ public final class CharArrayMemory implements Memory {
 
     public char[] getArray() {
         return array;
-    }
-
-    @Override
-    public int load8(long index, ReadAccessMode mode) {
-        throw new InvalidMemoryAccessException();
     }
 
     @Override
@@ -53,36 +44,6 @@ public final class CharArrayMemory implements Memory {
     }
 
     @Override
-    public int load32(long index, ReadAccessMode mode) {
-        throw new InvalidMemoryAccessException();
-    }
-
-    @Override
-    public long load64(long index, ReadAccessMode mode) {
-        throw new InvalidMemoryAccessException();
-    }
-
-    @Override
-    public VmObject loadRef(long index, ReadAccessMode mode) {
-        throw new InvalidMemoryAccessException();
-    }
-
-    @Override
-    public ValueType loadType(long index, ReadAccessMode mode) {
-        throw new InvalidMemoryAccessException();
-    }
-
-    @Override
-    public Pointer loadPointer(long index, ReadAccessMode mode) {
-        throw new InvalidMemoryAccessException();
-    }
-
-    @Override
-    public void store8(long index, int value, WriteAccessMode mode) {
-        throw new InvalidMemoryAccessException();
-    }
-
-    @Override
     public void store16(long index, int value, WriteAccessMode mode) {
         if (GlobalPlain.includes(mode)) {
             array[Math.toIntExact(index >>> 1)] = (char) value;
@@ -96,31 +57,6 @@ public final class CharArrayMemory implements Memory {
     }
 
     @Override
-    public void store32(long index, int value, WriteAccessMode mode) {
-        throw new InvalidMemoryAccessException();
-    }
-
-    @Override
-    public void store64(long index, long value, WriteAccessMode mode) {
-        throw new InvalidMemoryAccessException();
-    }
-
-    @Override
-    public void storeRef(long index, VmObject value, WriteAccessMode mode) {
-        throw new InvalidMemoryAccessException();
-    }
-
-    @Override
-    public void storeType(long index, ValueType value, WriteAccessMode mode) {
-        throw new InvalidMemoryAccessException();
-    }
-
-    @Override
-    public void storePointer(long index, Pointer value, WriteAccessMode mode) {
-        throw new InvalidMemoryAccessException();
-    }
-
-    @Override
     public void storeMemory(long destIndex, Memory src, long srcIndex, long size) {
         throw new UnsupportedOperationException();
     }
@@ -128,11 +64,6 @@ public final class CharArrayMemory implements Memory {
     @Override
     public void storeMemory(long destIndex, byte[] src, int srcIndex, int size) {
         throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public int compareAndExchange8(long index, int expect, int update, ReadAccessMode readMode, WriteAccessMode writeMode) {
-        throw new InvalidMemoryAccessException();
     }
 
     @Override
@@ -150,31 +81,6 @@ public final class CharArrayMemory implements Memory {
         } else {
             return (int) h16.compareAndExchange(array, Math.toIntExact(index >>> 1), (char) expect, (char) update) & 0xffff;
         }
-    }
-
-    @Override
-    public int compareAndExchange32(long index, int expect, int update, ReadAccessMode readMode, WriteAccessMode writeMode) {
-        throw new InvalidMemoryAccessException();
-    }
-
-    @Override
-    public long compareAndExchange64(long index, long expect, long update, ReadAccessMode readMode, WriteAccessMode writeMode) {
-        throw new InvalidMemoryAccessException();
-    }
-
-    @Override
-    public VmObject compareAndExchangeRef(long index, VmObject expect, VmObject update, ReadAccessMode readMode, WriteAccessMode writeMode) {
-        throw new InvalidMemoryAccessException();
-    }
-
-    @Override
-    public ValueType compareAndExchangeType(long index, ValueType expect, ValueType update, ReadAccessMode readMode, WriteAccessMode writeMode) {
-        throw new InvalidMemoryAccessException();
-    }
-
-    @Override
-    public Pointer compareAndExchangePointer(long index, Pointer expect, Pointer update, ReadAccessMode readMode, WriteAccessMode writeMode) {
-        throw new InvalidMemoryAccessException();
     }
 
     @Override
